+++ CHANGELOG.md
@@ -30,6 +30,8 @@
   short options; add `-k`, `-a`, `-c` options; remove the `--update-all` option;
   report total number of crates or number of outdated crates; remove examples;
   update dependencies
+* 0.13.0 (2023-12-13): Fix the `cargo install` command to update a crate; add
+  library docstrings; general cleanup
 
 [`atty`]: https://crates.io/crates/atty
 [`bunt`]: https://crates.io/crates/bunt